# [FIX] mail: improve performance of channel member_count